requirements.txt
pycryptodome==3.19.0
cryptography==41.0.7
pycryptodome==3.19.0